test: move outage backoff test to its own file
Kept it out of card.test.ts's monolith so a reader scanning gallery/
tests finds the end-to-end backoff coverage next to backoff.test.ts's
unit coverage, not buried in a 1700-line describe block.
