+++ test/card/gallery/backoff-integration.test.ts
@@ -1,0 +1,72 @@
+import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
+import { SolarViewCard } from "../../../src/card/card.js";
+import { urlCache } from "../../../src/card/gallery/url-cache.js";
+
+// End-to-end companion to backoff.test.ts's unit coverage: drives the real card, its
+// refresh_mins timer, and gallery-controller/resolve() plumbing through a sustained outage,
+// rather than calling Backoff or SourceResolver directly — confirms the cooldown actually
+// suppresses network calls at the boundary a user's browser would hit, not just at the
+// class-under-test boundary.
+describe("gallery backoff integration", () => {
+  beforeAll(() => {
+    if (!customElements.get("ha-planetary-solar-system-card-test")) {
+      customElements.define("ha-planetary-solar-system-card-test", SolarViewCard);
+    }
+  });
+
+  beforeEach(() => {
+    vi.stubGlobal("fetch", vi.fn().mockRejectedValue(new Error("fetch not stubbed for this test")));
+  });
+
+  afterEach(() => {
+    vi.unstubAllGlobals();
+    vi.useRealTimers();
+    urlCache.clear();
+  });
+
+  function createAndMount(config) {
+    const card = document.createElement("ha-planetary-solar-system-card-test");
+    card.setConfig({ gallery: { mode: "both" }, ...config });
+    document.body.appendChild(card);
+    return card;
+  }
+
+  it("a sustained sun-source outage backs off instead of retrying every refresh_mins tick", async () => {
+    // Real network attempts only, not decode-gate hits: every Image() construction is one
+    // real probe of a candidate URL (the primary guess plus, on failure, up to
+    // SUN_MAX_RETRIES fallback slots) — the number backoff exists to bound.
+    let imageAttempts = 0;
+    vi.stubGlobal(
+      "Image",
+      class {
+        src = "";
+        constructor() {
+          imageAttempts++;
+        }
+        decode() {
+          return Promise.reject(new Error("decode failed"));
+        }
+      }
+    );
+    vi.useFakeTimers();
+    const card = createAndMount({ refresh_mins: 1 });
+    await vi.advanceTimersByTimeAsync(0); // mount's own background fetch fails
+
+    const attemptsAfterMount = imageAttempts;
+    expect(attemptsAfterMount).toBeGreaterThan(0);
+
+    // 6 hours of 1-minute ticks (360 of them). Without backoff, each tick re-attempts
+    // (primary + retries) — hundreds of probes. With it, cooldown skips ticks that land
+    // inside the current backoff window, so this stays a small, capped number.
+    await vi.advanceTimersByTimeAsync(6 * 3600000);
+
+    expect(imageAttempts).toBeGreaterThan(attemptsAfterMount); // it did keep trying...
+    expect(imageAttempts).toBeLessThan(50); // ...but nowhere near the ~1440 a naive
+    // per-tick retry (primary + SUN_MAX_RETRIES fallbacks, every one of the 360 ticks)
+    // would have produced
+    expect(urlCache.inCooldown("sun")).toBe(true); // still backed off at the end of the outage
+
+    card.remove();
+    vi.useRealTimers();
+  });
+});
